server/requirements.txt
Flask==1.0.2
gevent_socketio==0.3.6
gunicorn==18.0.0
